feat(nestjs): Add alias `@SentryExceptionCaptured` for `@WithSentry`
