--- packages/nestjs/src/decorators/sentry-cron.ts
@@ -1,24 +1,0 @@
-import * as Sentry from '@sentry/node';
-import type { MonitorConfig } from '@sentry/types';
-
-/**
- * A decorator wrapping the native nest Cron decorator, sending check-ins to Sentry.
- */
-export const SentryCron = (monitorSlug: string, monitorConfig?: MonitorConfig): MethodDecorator => {
-  return (target: unknown, propertyKey, descriptor: PropertyDescriptor) => {
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    const originalMethod = descriptor.value as (...args: any[]) => Promise<any>;
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    descriptor.value = function (...args: any[]) {
-      return Sentry.withMonitor(
-        monitorSlug,
-        () => {
-          return originalMethod.apply(this, args);
-        },
-        monitorConfig,
-      );
-    };
-    return descriptor;
-  };
-};
