+++ packages/nestjs/src/index.ts
@@ -2,6 +2,9 @@ export * from '@sentry/node';
 
 export { init } from './sdk';
 
-export { SentryTraced } from './decorators/sentry-traced';
-export { SentryCron } from './decorators/sentry-cron';
-export { WithSentry } from './decorators/with-sentry';
+export {
+  SentryTraced,
+  SentryCron,
+  WithSentry,
+  SentryExceptionCaptured,
+} from './decorators';
